Task 221: Done
Fix length de las entidades
Reestructuración de paquetes de proyecto
Money en PatronDashboard
Importación de clase Inventor en Item y Toolkit
Actualización de modelo UML
