+++ src/main/java/acme/entities/Chirp.java
@@ -35,15 +35,15 @@ public class Chirp extends AbstractEntity{
 	protected Date moment;
 
 	@NotBlank
-	@Length(max = 101)
+	@Length(max = 100)
 	protected String title;
 
 	@NotBlank
-	@Length(max = 101)
+	@Length(max = 100)
 	protected String author;
 
 	@NotBlank
-	@Length(max = 256)
+	@Length(max = 255)
 	protected String body;
 
 	@Email
