+++ src/main/java/acme/entities/toolkits/Toolkit.java
@@ -1,11 +1,8 @@
-package acme.entities;
+package acme.entities.toolkits;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
-import javax.persistence.EnumType;
-import javax.persistence.Enumerated;
 import javax.persistence.ManyToOne;
-import javax.validation.Valid;
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Pattern;
@@ -13,61 +10,53 @@ import javax.validation.constraints.Pattern;
 import org.hibernate.validator.constraints.Length;
 import org.hibernate.validator.constraints.URL;
 
-import acme.framework.datatypes.Money;
 import acme.framework.entities.AbstractEntity;
+import acme.roles.Inventor;
 import lombok.Getter;
 import lombok.Setter;
 
 @Entity
 @Getter
 @Setter
-public class Item extends AbstractEntity{
-
+public class Toolkit extends AbstractEntity{
+	
 	// Serialisation identifier -----------------------------------------------
 
-	protected static final long	serialVersionUID	= 1L;
+	protected static final long serialVersionUID = 1L;
 	
 	// Attributes -------------------------------------------------------------
-
-	@Length(min=1,max=100)
-	@NotBlank
-	protected String name;
 	
-	@Pattern(regexp = "[A-Z]{3}-[0-9]{3}(-[A-Z])?$")
+	@Column(unique=true)
 	@NotNull
-	@Column(unique = true)
-	protected String code;
+	@Pattern(regexp="^[A-Z]{3}-[0-9]{3}(-[A-Z])?$")	
+	protected String 		code;
 	
-	@NotNull
-	@Enumerated(EnumType.STRING)
-	protected ItemType itemType;
 	
 	@NotBlank
 	@Length(min=1,max=100)
-	protected String technology;
+	protected String 		title;
 	
 	@NotBlank
 	@Length(min=1,max=255)
-	protected String description;
+	protected String 		description;
 	
-	@Valid
-	@NotNull
-	protected  Money retailPrice;
+	@NotBlank
+	@Length(min=1,max=255)
+	protected String 		assemblyNotes;
 	
 	@ManyToOne(optional=false)
 	@NotNull
 	protected Inventor inventor;
 	
-	@URL
-	@Column(name = "further_information")
-	protected String info;
+	@URL	
+	protected String		info;
 
 	// Derived attributes -----------------------------------------------------
+	
 
-	
-	
 	// Relationships ----------------------------------------------------------
 
-
+	
+	
 	
 }
