+++ src/main/java/acme/roles/Inventor.java
@@ -34,11 +34,11 @@ public class Inventor extends UserRole {
 	// Attributes -------------------------------------------------------------
 
 	@NotBlank
-	@Length(max = 101)
+	@Length(max = 100)
 	protected String			company;
 
 	@NotBlank
-	@Length(max = 256)
+	@Length(max = 255)
 	protected String			statement;
 	
 	@URL
